refactor(storage): extract game-storage detect & mount into standalone plugin
Move the device-agnostic game-storage detect & mount feature out of the
DMI-gated apex plugin into a new standalone plugins/storage/ plugin so it
works on any device, not just the OneXPlayer Apex.

- New plugins/storage/: manifest, StorageBackend (ungated getStatus/
  detectDrives/mountDrive/setDriveAutoMount), full UI page, backend +
  app specs, README, LICENSE.
- lib/storage.{ts,test.ts} moved verbatim (git mv); self-contained now
  (Run/RunResult defined locally instead of importing from ./xhci).
- apex: removed the storage import block, resolveTargetUser, storageDeps,
  the three storage RPCs, storage from getStatus(), the Storage drive
  card + its handlers/interfaces, FaHardDrive, and lsblk/mount/findmnt
  perms; reverted the manifest description. Gamepad recovery / xhci,
  hid-oxp blacklist, fingerprint wake and the wake listener untouched.
