+++ plugins/apex/backend.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect, beforeEach, afterEach, mock } from "bun:test";
+import { describe, it, expect, beforeEach, mock } from "bun:test";
 import type { EmitPayload } from "@loadout/types";
 
 /**
@@ -47,30 +47,6 @@ const fingerprintStatusImpl = mock(async () => ({
 const applyFingerprintImpl = mock(async () => ({ success: true, rebootRequired: true, steps: [] }));
 const revertFingerprintImpl = mock(async () => ({ success: true, rebootRequired: true, steps: [] }));
 
-const driveFixture = {
-  path: "/dev/nvme1n1p1",
-  label: "Games",
-  uuid: "GAME-1",
-  fstype: "ext4",
-  size: 1024 ** 4,
-  mounted: false,
-  mountpoint: null as string | null,
-  suggestedMountpoint: "/run/media/deck/Games",
-  steamLibraryFound: false,
-  inFstab: false,
-};
-const storageStatusImpl = mock(async () => ({ drives: [{ ...driveFixture }] }));
-const detectCandidatesImpl = mock(async () => [
-  { path: "/dev/nvme1n1p1", label: "Games", uuid: "GAME-1", fstype: "ext4", size: 1024 ** 4 },
-]);
-const mountCandidateImpl = mock(async () => ({
-  success: true,
-  mountpoint: "/run/media/deck/Games",
-  steamLibraryFound: true,
-}));
-const persistFstabImpl = mock(async () => ({ success: true }));
-const unpersistFstabImpl = mock(async () => ({ success: true }));
-
 mock.module("@loadout/devices", () => ({
   isApex: async () => isApexResult,
 }));
@@ -86,13 +62,6 @@ mock.module("./lib/fingerprint", () => ({
   getStatus: fingerprintStatusImpl,
   apply: applyFingerprintImpl,
   revert: revertFingerprintImpl,
-}));
-mock.module("./lib/storage", () => ({
-  getStorageStatus: storageStatusImpl,
-  detectCandidates: detectCandidatesImpl,
-  mountCandidate: mountCandidateImpl,
-  persistFstab: persistFstabImpl,
-  unpersistFstab: unpersistFstabImpl,
 }));
 
 // In-memory plugin storage so settings persist within a test without touching
@@ -117,42 +86,7 @@ mock.module("@loadout/wake", () => ({
   startWakeListener: startWakeListenerImpl,
 }));
 
-import ApexBackend, { resolveTargetUser } from "./backend";
-
-describe("resolveTargetUser", () => {
-  const HOME = process.env.HOME;
-  const USER = process.env.USER;
-  afterEach(() => {
-    if (HOME === undefined) delete process.env.HOME;
-    else process.env.HOME = HOME;
-    if (USER === undefined) delete process.env.USER;
-    else process.env.USER = USER;
-  });
-
-  it("prefers the --user arg the system unit passes (space form)", () => {
-    expect(resolveTargetUser(["loadout", "--user", "deck"])).toBe("deck");
-  });
-
-  it("accepts the --user=NAME form", () => {
-    expect(resolveTargetUser(["loadout", "--user=alice"])).toBe("alice");
-  });
-
-  it("falls back to HOME's basename when there's no --user (root service)", () => {
-    process.env.HOME = "/home/deck";
-    expect(resolveTargetUser(["loadout"])).toBe("deck");
-  });
-
-  it("handles ostree-style /var/home/<user>", () => {
-    process.env.HOME = "/var/home/bazzite";
-    expect(resolveTargetUser(["loadout"])).toBe("bazzite");
-  });
-
-  it("never resolves to root from HOME=/root, using $USER instead", () => {
-    process.env.HOME = "/root";
-    process.env.USER = "deck";
-    expect(resolveTargetUser(["loadout"])).toBe("deck");
-  });
-});
+import ApexBackend from "./backend";
 
 function makeBackend() {
   const events: EmitPayload[] = [];
@@ -171,11 +105,6 @@ describe("Apex backend", () => {
     fingerprintStatusImpl.mockClear();
     applyFingerprintImpl.mockClear();
     revertFingerprintImpl.mockClear();
-    storageStatusImpl.mockClear();
-    detectCandidatesImpl.mockClear();
-    mountCandidateImpl.mockClear();
-    persistFstabImpl.mockClear();
-    unpersistFstabImpl.mockClear();
     storage = {};
     capturedOnResume = null;
     stopSpy.mockClear();
@@ -366,123 +295,5 @@ describe("Apex backend", () => {
     expect(res.unsupported).toBe(true);
     expect(res.success).toBe(false);
     expect(startWakeListenerImpl).not.toHaveBeenCalled();
-  });
-
-  // ---------- game-storage detect & mount ----------
-
-  it("includes storage in getStatus on Apex hardware", async () => {
-    const { backend } = makeBackend();
-    await backend.onLoad();
-
-    const status = await backend.getStatus();
-    expect(status.storage?.drives[0]?.uuid).toBe("GAME-1");
-    expect(storageStatusImpl).toHaveBeenCalledTimes(1);
-  });
-
-  it("detectDrives returns the scan and candidates on Apex", async () => {
-    const { backend } = makeBackend();
-    await backend.onLoad();
-
-    const res = await backend.detectDrives();
-    expect(res.unsupported).toBeUndefined();
-    expect(res.drives[0]?.uuid).toBe("GAME-1");
-    expect(res.candidates?.[0]?.uuid).toBe("GAME-1");
-    expect(detectCandidatesImpl).toHaveBeenCalledTimes(1);
-  });
-
-  it("detectDrives short-circuits on non-Apex hardware", async () => {
-    isApexResult = false;
-    const { backend } = makeBackend();
-    await backend.onLoad();
-
-    const res = await backend.detectDrives();
-    expect(res.unsupported).toBe(true);
-    expect(res.drives).toEqual([]);
-    expect(detectCandidatesImpl).not.toHaveBeenCalled();
-  });
-
-  it("mountDrive mounts through the lib and emits statusChanged", async () => {
-    const { backend, events } = makeBackend();
-    await backend.onLoad();
-
-    const res = await backend.mountDrive("GAME-1");
-    expect(res.success).toBe(true);
-    expect(res.mountpoint).toBe("/run/media/deck/Games");
-    expect(res.steamLibraryFound).toBe(true);
-    expect(mountCandidateImpl).toHaveBeenCalledWith(expect.anything(), { uuid: "GAME-1" });
-    expect(events).toEqual([{ event: "statusChanged", data: undefined }]);
-  });
-
-  it("refuses to mount on non-Apex hardware", async () => {
-    isApexResult = false;
-    const { backend } = makeBackend();
-    await backend.onLoad();
-
-    const res = await backend.mountDrive("GAME-1");
-    expect(res.unsupported).toBe(true);
-    expect(res.success).toBe(false);
-    expect(mountCandidateImpl).not.toHaveBeenCalled();
-  });
-
-  it("setDriveAutoMount(true) persists the resolved mount point", async () => {
-    const { backend, events } = makeBackend();
-    await backend.onLoad();
-
-    const res = await backend.setDriveAutoMount("GAME-1", true);
-    expect(res.success).toBe(true);
-    expect(persistFstabImpl).toHaveBeenCalledWith(expect.anything(), {
-      uuid: "GAME-1",
-      mountpoint: "/run/media/deck/Games",
-      fstype: "ext4",
-    });
-    expect(unpersistFstabImpl).not.toHaveBeenCalled();
-    expect(events).toEqual([{ event: "statusChanged", data: undefined }]);
-  });
-
-  it("setDriveAutoMount(true) persists the live mount point when already mounted", async () => {
-    storageStatusImpl.mockImplementationOnce(async () => ({
-      drives: [{ ...driveFixture, mounted: true, mountpoint: "/run/media/deck/Games" }],
-    }));
-    const { backend } = makeBackend();
-    await backend.onLoad();
-
-    await backend.setDriveAutoMount("GAME-1", true);
-    expect(persistFstabImpl).toHaveBeenCalledWith(expect.anything(), {
-      uuid: "GAME-1",
-      mountpoint: "/run/media/deck/Games",
-      fstype: "ext4",
-    });
-  });
-
-  it("setDriveAutoMount(false) removes the fstab entry without needing the drive", async () => {
-    const { backend } = makeBackend();
-    await backend.onLoad();
-
-    const res = await backend.setDriveAutoMount("GAME-1", false);
-    expect(res.success).toBe(true);
-    expect(unpersistFstabImpl).toHaveBeenCalledWith(expect.anything(), { uuid: "GAME-1" });
-    expect(persistFstabImpl).not.toHaveBeenCalled();
-  });
-
-  it("setDriveAutoMount errors when enabling for an unknown drive", async () => {
-    storageStatusImpl.mockImplementationOnce(async () => ({ drives: [] }));
-    const { backend } = makeBackend();
-    await backend.onLoad();
-
-    const res = await backend.setDriveAutoMount("MISSING", true);
-    expect(res.success).toBe(false);
-    expect(persistFstabImpl).not.toHaveBeenCalled();
-  });
-
-  it("refuses to set auto-mount on non-Apex hardware", async () => {
-    isApexResult = false;
-    const { backend } = makeBackend();
-    await backend.onLoad();
-
-    const res = await backend.setDriveAutoMount("GAME-1", true);
-    expect(res.unsupported).toBe(true);
-    expect(res.success).toBe(false);
-    expect(persistFstabImpl).not.toHaveBeenCalled();
-    expect(unpersistFstabImpl).not.toHaveBeenCalled();
   });
 });
