+++ plugins/storage/README.md
@@ -1,0 +1,12 @@
+# Storage
+
+> Detect and mount a game-storage drive (e.g. a second SSD holding a Steam library) that the system stopped auto-mounting after an update, and optionally pin it in /etc/fstab so it survives future updates. Works on any device.
+
+## Screenshots
+
+![Storage](./assets/screenshot.png)
+
+## See also
+
+- [All plugins](../../README.md#plugins)
+- [Plugin model](../../README.md#plugin-model)
