+++ plugins/storage/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "@loadout/plugin-storage-mount",
+  "version": "0.0.1",
+  "type": "module",
+  "dependencies": {
+    "react": "^18.3.0",
+    "react-dom": "^18.3.0",
+    "react-icons": "^5.6.0",
+    "@loadout/ui": "workspace:*"
+  },
+  "plugin": {
+    "id": "storage",
+    "name": "Storage",
+    "description": "Detect and mount a game-storage drive (e.g. a second SSD holding a Steam library) that the system stopped auto-mounting after an update, and optionally pin it in /etc/fstab so it survives future updates. Works on any device.",
+    "permissions": {
+      "commands": [
+        "lsblk",
+        "mount",
+        "findmnt"
+      ]
+    },
+    "category": "Device",
+    "subtitle": "Detect & mount game drives",
+    "target": {
+      "type": "overlay"
+    }
+  }
+}
